refactor(eventviewerx): 🔨 extract restored PowerShell script
